Reorganizing the ECDSA files
