+++ crates/threshold-signatures/src/ecdsa.rs
@@ -3,23 +3,13 @@
 pub mod ot_based_ecdsa;
 pub mod robust_ecdsa;
 
-use hkdf::Hkdf;
+mod rerandomization;
+mod signature;
 
-use elliptic_curve::{
-    bigint::U256,
-    ops::{Invert, Reduce},
-    point::AffineCoordinates,
-    scalar::IsHigh,
-    sec1::ToEncodedPoint,
-    PrimeField,
-};
-
-use frost_secp256k1::{Field, Group, Secp256K1Group, Secp256K1ScalarField};
-use k256::{AffinePoint, ProjectivePoint};
+pub(crate) use frost_secp256k1::{Field, Group, Secp256K1Group, Secp256K1ScalarField};
+pub(crate) use k256::{AffinePoint, ProjectivePoint};
 
 use crate::crypto::ciphersuite::{BytesOrder, Ciphersuite, ScalarSerializationFormat};
-use crate::errors::ProtocolError;
-use crate::participants::ParticipantList;
 
 pub use frost_secp256k1::Secp256K1Sha256;
 pub type KeygenOutput = crate::KeygenOutput<Secp256K1Sha256>;
@@ -30,6 +20,10 @@ pub type CoefficientCommitment = frost_core::keys::CoefficientCommitment<Secp256
 pub type Polynomial = crate::crypto::polynomials::Polynomial<Secp256K1Sha256>;
 pub type PolynomialCommitment = crate::crypto::polynomials::PolynomialCommitment<Secp256K1Sha256>;
 
+pub use rerandomization::RerandomizationArguments;
+pub(crate) use signature::x_coordinate;
+pub use signature::{Signature, SignatureOption};
+
 impl ScalarSerializationFormat for Secp256K1Sha256 {
     fn bytes_order() -> BytesOrder {
         BytesOrder::BigEndian
@@ -38,214 +32,16 @@ impl ScalarSerializationFormat for Secp256K1Sha256 {
 
 impl Ciphersuite for Secp256K1Sha256 {}
 
-/// Get the x coordinate of a point, as a scalar
-pub(crate) fn x_coordinate(point: &AffinePoint) -> Scalar {
-    <Scalar as Reduce<U256>>::reduce_bytes(&point.x())
-}
-
-/// Represents a signature that supports different variants of ECDSA.
-///
-/// An ECDSA signature is usually two scalars.
-/// The first is derived from using the x-coordinate of an elliptic curve point (`big_r`),
-/// and the second is computed using the typical ecdsa signing equation.
-/// Deriving the x-coordination implies losing information about `big_r`, some variants
-/// may thus include an extra information to recover this point.
-///
-/// This signature supports all variants by containing `big_r` entirely
-#[derive(Debug, Clone)]
-pub struct Signature {
-    /// This is the entire first point.
-    pub big_r: AffinePoint,
-    /// This is the second scalar, normalized to be in the lower range.
-    pub s: Scalar,
-}
-
-impl Signature {
-    // This verification tests the signature including whether s has been normalized
-    pub fn verify(&self, public_key: &AffinePoint, msg_hash: &Scalar) -> bool {
-        let r: Scalar = x_coordinate(&self.big_r);
-        if r.is_zero().into() || self.s.is_zero().into() {
-            return false;
-        }
-        // Check if s has been normalized
-        if self.s.is_high().into() {
-            return false;
-        }
-        // tested earlier is not zero, so inversion will not raise an error and unwrap cannot panic
-        let s_inv = self.s.invert_vartime().unwrap();
-        let reproduced = (ProjectivePoint::GENERATOR * (*msg_hash * s_inv))
-            + (ProjectivePoint::from(*public_key) * (r * s_inv));
-        x_coordinate(&reproduced.into()) == r
-    }
-}
-
-/// None for participants and Some for coordinator
-pub type SignatureOption = Option<Signature>;
-
-/// The arguments used to derive randomness used for presignature rerandomization.
-/// Presignature rerandomization has been thoroughly described in
-/// \[GS21\] <https://eprint.iacr.org/2021/1330.pdf>
-///
-/// *** Warning ***
-/// Following \[GS21\] <https://eprint.iacr.org/2021/1330.pdf>, the entropy should
-/// be public, freshly generated, and unpredictable.
-// Cannot derive Debug here because an external type inside Tweak does not implement it
-#[derive(Clone)]
-pub struct RerandomizationArguments {
-    // Preferable (but non-binding) the master public key
-    pub pk: AffinePoint,
-    pub tweak: Tweak,
-    pub msg_hash: [u8; 32],
-    pub big_r: AffinePoint,
-    pub participants: ParticipantList,
-    /// Fresh, Unpredictable, and Public source of entropy
-    pub entropy: [u8; 32],
-}
-
-impl RerandomizationArguments {
-    /// The following salt is picked by hashing with sha256
-    /// "NEAR 6.4478$ 7:20pm CEST 2024-11-24"
-    /// Based on [Krawczyk10] paper:
-    /// ``[...] in most applications the extractor key (or salt) can be used
-    /// repeatedly with many (independent) samples from the same source [...]``
-    const SALT: [u8; 32] = [
-        0x32, 0x8a, 0x47, 0xc2, 0xb8, 0x79, 0x44, 0x45, 0x25, 0x5c, 0x16, 0x47, 0x60, 0x8d, 0xf5,
-        0xdb, 0x85, 0xc6, 0x8b, 0xb0, 0xe7, 0x17, 0x0a, 0xbe, 0xc5, 0x34, 0xdf, 0x27, 0x64, 0xa4,
-        0x58, 0x31,
-    ];
-
-    pub fn new(
-        pk: AffinePoint,
-        tweak: Tweak,
-        msg_hash: [u8; 32],
-        big_r: AffinePoint,
-        participants: ParticipantList,
-        entropy: [u8; 32],
-    ) -> Self {
-        Self {
-            pk,
-            tweak,
-            msg_hash,
-            big_r,
-            participants,
-            entropy,
-        }
-    }
-
-    /// Derives a random string from the public key, tweak, message hash, presignature R,
-    /// set of participants and the entropy.
-    ///
-    /// Outputs a random string computed as HKDF(entropy, pk, hash, R, participants)
-    pub fn derive_randomness(&self) -> Result<Scalar, ProtocolError> {
-        // create a string containing (pk, msg_hash, big_r, sorted(participants))
-        let pk_encoded_point = self.pk.to_encoded_point(true);
-        let encoded_pk: &[u8] = pk_encoded_point.as_bytes();
-        let encoded_tweak: &[u8] = &<Secp256K1ScalarField as Field>::serialize(&self.tweak.value());
-        let encoded_msg_hash: &[u8] = &self.msg_hash;
-        let big_r_encoded_point = self.big_r.to_encoded_point(true);
-        let encoded_big_r: &[u8] = big_r_encoded_point.as_bytes();
-
-        // concatenate all the bytes
-        let mut concatenation = Vec::new();
-        // 1 byte counter, used in the unlikely case that the hash result is 0
-        concatenation.extend_from_slice(&[0u8, 1]);
-        concatenation.extend_from_slice(encoded_pk);
-        concatenation.extend_from_slice(encoded_tweak);
-        concatenation.extend_from_slice(encoded_msg_hash);
-        concatenation.extend_from_slice(encoded_big_r);
-        // Append each ParticipantId's
-        for participant in self.participants.participants() {
-            concatenation.extend_from_slice(&participant.bytes());
-        }
-
-        // initiate hkdf with the salt and with some `good' entropy
-        let hk = Hkdf::<sha3::Sha3_256>::new(Some(&Self::SALT), &self.entropy);
-
-        let mut delta = Scalar::ZERO;
-        // If the randomness created is 0 then we want to generate a new randomness
-        while bool::from(delta.is_zero()) {
-            // Generate randomization out of HKDF(counter, entropy, pk, msg_hash, big_r, participants, )
-            // where entropy is a public but unpredictable random string.
-            // The counter depends on the number of times we enter into this loop
-            let mut okm = [0u8; 32];
-
-            hk.expand(&concatenation, &mut okm)
-                .map_err(|_| ProtocolError::HashingError)?;
-
-            // derive the randomness delta
-            delta = Scalar::from_repr(okm.into()).unwrap_or(
-                // if delta falls outside the field
-                // probability is negligible: in the order of 1/2^224
-                Scalar::ZERO,
-            );
-            // Increment the counter, the probability that this overflows is astronomically low
-            let concatenation_0 = concatenation
-                .first_mut()
-                .ok_or(ProtocolError::InvalidIndex)?;
-            *concatenation_0 += 1;
-        }
-        Ok(delta)
-    }
-}
-
 #[cfg(test)]
 mod test {
     use crate::{
-        ecdsa::{
-            KeygenOutput, RerandomizationArguments, Scalar, Secp256K1Sha256, Signature, Tweak,
-        },
-        participants::ParticipantList,
-        test_utils::{
-            ecdsa_generate_rerandpresig_args, generate_participants,
-            generate_participants_with_random_ids, MockCryptoRng,
-        },
+        ecdsa::{KeygenOutput, Scalar, Secp256K1Sha256},
+        test_utils::{generate_participants, MockCryptoRng},
     };
 
-    use elliptic_curve::ops::{Invert, LinearCombination, Reduce};
-
-    use frost_core::{keys::SigningShare, Ciphersuite, SigningKey as FrostSigningKey};
-
-    use k256::{
-        ecdsa::{signature::Verifier, SigningKey},
-        ProjectivePoint, Secp256k1,
-    };
+    use frost_core::{keys::SigningShare, SigningKey as FrostSigningKey};
     use rand::SeedableRng;
-    use rand_core::{CryptoRngCore, RngCore};
-    use sha2::{digest::FixedOutput, Digest, Sha256};
     type C = Secp256K1Sha256;
-
-    #[test]
-    fn test_verify() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let msg = b"Hello from Near";
-        let mut hasher = Sha256::new();
-        hasher.update(msg);
-
-        let sk = SigningKey::random(&mut rng);
-        let pk = ecdsa::VerifyingKey::from(&sk);
-        let (sig, _) = sk.sign_digest_recoverable(hasher.clone()).unwrap();
-        assert!(pk.verify(msg, &sig).is_ok());
-
-        let z_bytes = hasher.clone().finalize_fixed();
-        let z =
-            <Scalar as Reduce<<Secp256k1 as elliptic_curve::Curve>::Uint>>::reduce_bytes(&z_bytes);
-        let (r, s) = sig.split_scalars();
-        let s_inv = *s.invert_vartime();
-        let u1 = z * s_inv;
-        let u2 = *r * s_inv;
-        let pk = ProjectivePoint::from(pk.as_affine());
-        let big_r =
-            ProjectivePoint::lincomb(&ProjectivePoint::GENERATOR, &u1, &pk, &u2).to_affine();
-
-        let full_sig = Signature {
-            big_r,
-            s: *s.as_ref(),
-        };
-
-        let is_verified = full_sig.verify(&pk.to_affine(), &z);
-        // Should always be ok as signature contains Uint i.e. normalized elements
-        assert!(is_verified);
-    }
 
     #[test]
     fn keygen_output_should_be_serializable() {
@@ -267,109 +63,6 @@ mod test {
             serialized_keygen_output,
             "{\"private_share\":\"0000000000000000000000000000000000000000000000000000000000000001\",\"public_key\":\"0351177dde89242d9121d787a681bd2a0bd6013428a6b83e684a253815db96d8b3\"}"
         );
-    }
-
-    // Outputs pk, R, hash, participants, entropy, randomness
-    fn compute_random_outputs(
-        rng: &mut impl CryptoRngCore,
-        num_participants: usize,
-    ) -> (RerandomizationArguments, Scalar) {
-        let (_, big_r) = <C>::generate_nonce(rng);
-        let (_, pk) = <C>::generate_nonce(rng);
-        let pk = frost_core::VerifyingKey::new(pk);
-        let big_r = big_r.to_affine();
-
-        // Generate unique ten ParticipantId values
-        let participants = generate_participants_with_random_ids(num_participants, rng);
-        // Generate Rerandomization arguments
-        let (args, _) = ecdsa_generate_rerandpresig_args(rng, &participants, pk, big_r);
-
-        let delta = args.derive_randomness().unwrap();
-        (args, delta)
-    }
-
-    #[test]
-    fn test_different_pk() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-        // different pk
-        let (_, pk) = <C>::generate_nonce(&mut rng);
-        args.pk = pk.to_affine();
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    #[test]
-    fn test_different_tweak() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-        // different pk
-        args.tweak = Tweak::new(frost_core::random_nonzero::<Secp256K1Sha256, _>(&mut rng));
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    #[test]
-    fn test_different_msg_hash() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-        // different msg_hash
-        args.msg_hash = [0; 32];
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    #[test]
-    fn test_different_big_r() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-        // different big_r
-        let (_, big_r) = <C>::generate_nonce(&mut rng);
-        args.big_r = big_r.to_affine();
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    #[test]
-    fn test_different_participants() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-        // different participants set
-        let participants = generate_participants_with_random_ids(num_participants, &mut rng);
-        args.participants = ParticipantList::new(&participants).unwrap();
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    #[test]
-    fn test_different_entropy() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-
-        // different entropy
-        rng.fill_bytes(&mut args.entropy);
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_ne!(delta, delta_prime);
-    }
-
-    // Test that with different order of participants, the randomness is the same.
-    #[test]
-    fn test_same_randomness() {
-        let mut rng = MockCryptoRng::seed_from_u64(42);
-        let num_participants = 10;
-        let (mut args, delta) = compute_random_outputs(&mut rng, num_participants);
-
-        // reshuffle
-        args.participants = args.participants.shuffle(rng).unwrap();
-        let delta_prime = args.derive_randomness().unwrap();
-        assert_eq!(delta, delta_prime);
     }
 
     #[test]
