+++ crates/threshold-signatures/src/ecdsa/robust_ecdsa/presign.rs
@@ -1,9 +1,8 @@
-use super::{PresignArguments, PresignOutput};
 use crate::participants::{Participant, ParticipantList, ParticipantMap};
 use crate::{
     ecdsa::{
-        CoefficientCommitment, Field, Polynomial, PolynomialCommitment, Scalar,
-        Secp256K1ScalarField, Secp256K1Sha256,
+        AffinePoint, CoefficientCommitment, Field, KeygenOutput, Polynomial, PolynomialCommitment,
+        RerandomizationArguments, Scalar, Secp256K1ScalarField, Secp256K1Sha256,
     },
     errors::{InitializationError, ProtocolError},
     protocol::{
@@ -11,14 +10,130 @@ use crate::{
         internal::{make_protocol, Comms, SharedChannel},
         Protocol,
     },
-    SigningShare,
+    MaxMalicious, SigningShare,
 };
 use frost_core::serialization::SerializableScalar;
 use frost_secp256k1::{Group, Secp256K1Group};
 use rand_core::CryptoRngCore;
-use subtle::ConstantTimeEq;
+use serde::{Deserialize, Serialize};
+use subtle::{Choice, ConstantTimeEq};
+use zeroize::ZeroizeOnDrop;
 
 type C = Secp256K1Sha256;
+
+/// The necessary inputs for the creation of a presignature.
+pub struct PresignArguments {
+    /// The output of key generation, i.e. our share of the secret key, and the public key package.
+    /// This is of type `KeygenOutput<Secp256K1Sha256>` from Frost implementation
+    pub keygen_out: KeygenOutput,
+    /// The desired threshold for the presignature, which must match the original threshold
+    pub max_malicious: MaxMalicious,
+}
+
+/// The output of the presigning protocol.
+/// Contains the signature precomputed elements
+/// independently of the message
+#[derive(Clone, Serialize, Deserialize, ZeroizeOnDrop)]
+pub struct PresignOutput {
+    /// The public nonce commitment.
+    #[zeroize(skip)]
+    pub big_r: AffinePoint,
+
+    /// Our secret shares of the nonces.
+    pub c: Scalar,
+    pub e: Scalar,
+    pub alpha: Scalar,
+    pub beta: Scalar,
+}
+
+impl_secret_debug!(PresignOutput {
+    show: [big_r],
+    redact: [c, e, alpha, beta]
+});
+
+impl ConstantTimeEq for PresignOutput {
+    fn ct_eq(&self, other: &Self) -> Choice {
+        self.big_r.ct_eq(&other.big_r)
+            & self.c.ct_eq(&other.c)
+            & self.e.ct_eq(&other.e)
+            & self.alpha.ct_eq(&other.alpha)
+            & self.beta.ct_eq(&other.beta)
+    }
+}
+
+impl PartialEq for PresignOutput {
+    fn eq(&self, other: &Self) -> bool {
+        self.ct_eq(other).into()
+    }
+}
+impl Eq for PresignOutput {}
+
+/// The output of the presigning protocol.
+/// Contains the signature precomputed elements
+/// independently of the message
+#[derive(Clone, Serialize, Deserialize, ZeroizeOnDrop)]
+pub struct RerandomizedPresignOutput {
+    /// The rerandomized public nonce commitment.
+    #[zeroize(skip)]
+    pub(super) big_r: AffinePoint,
+
+    /// Our rerandomized secret shares of the nonces.
+    pub(super) e: Scalar,
+    pub(super) alpha: Scalar,
+    pub(super) beta: Scalar,
+}
+
+impl_secret_debug!(RerandomizedPresignOutput {
+    show: [big_r],
+    redact: [e, alpha, beta]
+});
+
+impl RerandomizedPresignOutput {
+    pub fn rerandomize_presign(
+        presignature: &PresignOutput,
+        args: &RerandomizationArguments,
+    ) -> Result<Self, ProtocolError> {
+        if presignature.big_r != args.big_r {
+            return Err(ProtocolError::IncompatibleRerandomizationInputs);
+        }
+        let delta = args.derive_randomness()?;
+        if delta.is_zero().into() {
+            return Err(ProtocolError::ZeroScalar);
+        }
+
+        // cannot be zero due to the previous check
+        let inv_delta = delta.invert().unwrap();
+
+        // delta * R
+        let rerandomized_big_r = presignature.big_r * delta;
+
+        // alpha * delta^{-1}
+        let rerandomized_alpha = presignature.alpha * inv_delta;
+
+        // (beta + c*tweak) * delta^{-1}
+        let rerandomized_beta =
+            (presignature.beta + presignature.c * args.tweak.value()) * inv_delta;
+
+        Ok(Self {
+            big_r: rerandomized_big_r.into(),
+            alpha: rerandomized_alpha,
+            beta: rerandomized_beta,
+            e: presignature.e,
+        })
+    }
+
+    #[cfg(test)]
+    /// Outputs the same elements as in the `PresignatureOutput`
+    /// Used for testing the core schemes without rerandomization
+    pub fn new_without_rerandomization(presignature: &PresignOutput) -> Self {
+        Self {
+            big_r: presignature.big_r,
+            alpha: presignature.alpha,
+            beta: presignature.beta,
+            e: presignature.e,
+        }
+    }
+}
 
 /// Maximum incoming buffer entries for the robust ECDSA presign protocol.
 pub(crate) const ROBUST_ECDSA_PRESIGN_MAX_INCOMING_BUFFER_ENTRIES: usize = 3;
