+++ crates/threshold-signatures/src/ecdsa/robust_ecdsa/test.rs
@@ -5,10 +5,9 @@ use super::{presign::presign, sign::sign, PresignArguments, PresignOutput};
 use crate::crypto::hash::test::scalar_hash_secp256k1;
 use crate::ecdsa::robust_ecdsa::RerandomizedPresignOutput;
 use crate::ecdsa::{
-    Element, ParticipantList, RerandomizationArguments, Secp256K1Sha256, Signature,
-    SignatureOption, Tweak,
+    Element, RerandomizationArguments, Secp256K1Sha256, Signature, SignatureOption, Tweak,
 };
-use crate::participants::Participant;
+use crate::participants::{Participant, ParticipantList};
 use crate::protocol::Protocol;
 use crate::test_utils::{
     assert_public_key_invariant, check_one_coordinator_output, generate_participants,
